Removal of NotificationEvent Request, Response , Search Results
Addition of NotificationEventDoc and Removal of NotificationEventInfo

Addition of NotificationEventDocTests
